[sw,cryptolib] Only return input errors on user errors
The error reported in case of the is on curve check should
only be an input error when the initial public key check is
executed. After the initial check, the is on curve check would
only return an error in case of an FI attack. In that case we
should not report the error as a user error.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>