[keymgr/keymgr_dpe,rtl] Fix wrong cast
`rom_digest_i[k].data` is casted to `MaxWidth`. In SV, the cast fills
up the MSBs with 0. This means that an all-1 vector is no  longer an
all-1 vector after the cast. E.g., when casting 8'hFF to 16-bits, we
would get 16'h00ff.

This is problematic as we use the casted value in the `valid_chk`
function. This function checks if the value is an all-1 or all-0
vector.

So far this was not problematic as `MaxWidth` and `rom_digest_i[k].data`
are both 256-bits. However, we should still fix this by using
`prim_msb_extend`.

This issue has been flagged by Muhammad Hafiz Abu Hassan from the Malaysian
Cryptology Technology and Management Centre during the HACK@CHES'25 event.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>